requirements.txt
numpy
scipy
scikit-learn

dwave-system
dwave-qbsolv
dwave-neal
amplify
matplotlib